+++ src/main/java/mju/scholarship/scholoarship/Scholarship.java
@@ -35,7 +35,7 @@ public class Scholarship {
     private String province;
     private String city;
     private String department; // major -> 전공 (?)
-    private Integer grade; // 1.0 - 4.5
+    private Double grade; // 1.0 - 4.5
     private Integer incomeQuantile; // 1-10
 
     @Builder
@@ -50,7 +50,7 @@ public class Scholarship {
                        String province,
                        String city,
                        String department,
-                       Integer grade,
+                       Double grade,
                        Integer incomeQuantile) {
 
         this.price = price;
